serp_analysis/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
google-cloud-bigquery==3.3.0
requests==2.28.1
google-cloud-secret-manager==2.12.0